Switched ActionListActivity to use a CursorAdapter instead of an ArrayAdapter. Realized the design of DatabaseHelper, ZoneDatabase and ActionDatabase is pretty lame, trying to work through it.

+++ src/org/anhonesteffort/polygons/database/ZoneDatabase.java
@@ -15,6 +15,16 @@ public class ZoneDatabase {
 
   private static final String TAG = "ZoneDatabase";
 
+  protected static final String ZONE_TABLE = "zone";
+  protected static final String ZONE_ID    = "_id";
+  protected static final String ZONE_LABEL = "label";
+
+  protected static final String POINT_TABLE     = "point";
+  protected static final String POINT_ID        = "_id";
+  protected static final String POINT_ZONE_ID   = "zone_id";
+  protected static final String POINT_LATITUDE  = "latitude";
+  protected static final String POINT_LONGITUDE = "longitude";
+
   private DatabaseHelper dbHelper;
   private ArrayList<GeometryChangeListener> listeners = new ArrayList<GeometryChangeListener>();
 
@@ -34,29 +44,29 @@ public class ZoneDatabase {
     dbHelper.exec("DROP TABLE IF EXISTS selected");
 
     dbHelper.exec("CREATE TABLE IF NOT EXISTS zone (" +
-                    "id INTEGER NOT NULL PRIMARY KEY, " +
+                    "_id INTEGER NOT NULL PRIMARY KEY, " +
                     "label VARCHAR(100) NOT NULL" +
                   ")");
     dbHelper.exec("SELECT AddGeometryColumn('zone', 'geometry', 4326, 'GEOMETRY')");
 
     dbHelper.exec("CREATE TABLE IF NOT EXISTS point (" +
-                    "id INTEGER NOT NULL PRIMARY KEY, " +
+                    "_id INTEGER NOT NULL PRIMARY KEY, " +
                     "zone_id INTEGER NOT NULL, " +
                     "latitude REAL NOT NULL, " +
                     "longitude REAL NOT NULL, " +
-                    "FOREIGN KEY (zone_id) REFERENCES zone(id) ON DELETE CASCADE" +
+                    "FOREIGN KEY (zone_id) REFERENCES zone(_id) ON DELETE CASCADE" +
                   ")");
 
     dbHelper.exec("CREATE TABLE IF NOT EXISTS occupy (" +
                     "zone_id INTEGER NOT NULL, " +
                     "PRIMARY KEY (zone_id), " +
-                    "FOREIGN KEY (zone_id) REFERENCES zone(id) ON DELETE CASCADE" +
+                    "FOREIGN KEY (zone_id) REFERENCES zone(_id) ON DELETE CASCADE" +
                   ")");
 
     dbHelper.exec("CREATE TABLE IF NOT EXISTS selected (" +
                     "zone_id INTEGER NOT NULL, " +
                     "PRIMARY KEY (zone_id), " +
-                    "FOREIGN KEY (zone_id) REFERENCES zone(id) ON DELETE CASCADE" +
+                    "FOREIGN KEY (zone_id) REFERENCES zone(_id) ON DELETE CASCADE" +
                   ")");
   }
 
@@ -85,15 +95,8 @@ public class ZoneDatabase {
     listeners.remove(listener);
   }
 
-  private void geometryChange() {
-    for(GeometryChangeListener listener : listeners) {
-      if(listener != null)
-      listener.onGeometryChange();
-    }
-  }
-
   public boolean isLabelAvailable(String label) {
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT id FROM zone WHERE label = '" +
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT _id FROM zone WHERE label = '" +
                                                     DatabaseHelper.escapeString(label) + "'");
 
     if(zoneRecords.getCount() > 0) {
@@ -105,7 +108,14 @@ public class ZoneDatabase {
     return true;
   }
 
-  public PointRecord addPoint(PointRecord point, int zone_id) {
+  private void geometryChange() {
+    for(GeometryChangeListener listener : listeners) {
+      if(listener != null)
+        listener.onGeometryChange();
+    }
+  }
+
+  private PointRecord addPoint(PointRecord point, int zone_id) {
     int point_id = -1;
     ContentValues values = new ContentValues();
 
@@ -114,40 +124,41 @@ public class ZoneDatabase {
     values.put("longitude", point.getX());
     dbHelper.insert("point", values);
 
-    SpatialCursor pointRecords = dbHelper.prepare("SELECT id FROM point ORDER BY id DESC LIMIT 1");
+    SpatialCursor pointRecords = dbHelper.prepare("SELECT _id FROM point ORDER BY _id DESC LIMIT 1");
     if(pointRecords.moveToNext())
       point_id = pointRecords.getInt(0);
 
     pointRecords.close();
-    return new PointRecord(point_id, point.getX(), point.getY());
-  }
-
-  public void removePoint(int point_id) {
-    dbHelper.exec("DELETE FROM point WHERE id = '" + point_id + "'");
-  }
-
-  public void updatePoint(PointRecord point, int zone_id) {
-    dbHelper.exec("UPDATE point SET " +
-                    "zone_id ='" + zone_id + "', " +
-                    "latitude = '" + point.getY() + "', " +
-                    "longitude = '" + point.getX() + "' " +
-                  "WHERE id = '" + point.getId() + "'");
+    return new PointRecord(point_id, zone_id, point.getX(), point.getY());
   }
 
   public PointRecord getPoint(int point_id) {
     PointRecord outPoint = null;
-    SpatialCursor pointRecords = dbHelper.prepare("SELECT id, longitude, latitude FROM point WHERE id = '" +
-                                                                                               point_id + "'");
+    SpatialCursor pointRecords = dbHelper.prepare("SELECT _id, zone_id, longitude, latitude FROM point WHERE _id = '" +
+                                                                                                         point_id + "'");
 
     if(pointRecords.moveToNext())
-      outPoint = new PointRecord(pointRecords.getInt(0), pointRecords.getDouble(1), pointRecords.getDouble(2));
+      outPoint = new PointRecord(pointRecords.getInt(0),
+                                 pointRecords.getInt(1),
+                                 pointRecords.getDouble(2),
+                                 pointRecords.getDouble(3));
 
     return outPoint;
   }
 
+  private boolean zoneExists(int zone_id) {
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT _id FROM zone WHERE _id = '" + zone_id + "'");
+    if(zoneRecords.moveToNext()) {
+      zoneRecords.close();
+      return true;
+    }
+
+    zoneRecords.close();
+    return false;
+  }
+
   public ZoneRecord addZone(String label) {
     Log.d(TAG, "addZone(), label: " + label);
-
     ZoneRecord outZone = new ZoneRecord(-1, label);
 
     ContentValues values = new ContentValues();
@@ -155,64 +166,55 @@ public class ZoneDatabase {
     values.put("geometry", "ST_GeomFromText('POINT(0.0 0.0)', 4326)");
     dbHelper.insert("zone", values);
 
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT id FROM zone ORDER BY id DESC LIMIT 1");
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT _id FROM zone ORDER BY _id DESC LIMIT 1");
     if(zoneRecords.moveToNext())
       outZone = new ZoneRecord(zoneRecords.getInt(0), label);
 
     zoneRecords.close();
+    dbHelper.getActionDatabase().initZoneActions(outZone.getId());
     return outZone;
   }
 
   public void deleteZone(int zone_id) {
     Log.d(TAG, "deleteZone(), id: " + zone_id);
-    dbHelper.exec("DELETE FROM zone WHERE id = '" + zone_id + "'");
+    dbHelper.exec("DELETE FROM zone WHERE _id = '" + zone_id + "'");
   }
 
   public ZoneRecord updateZone(ZoneRecord zone) {
     Log.d(TAG, "updateZone(), id: " + zone.getId() + ", label: " + zone.getLabel());
-    String sql;
 
-    // Enforce point limits.
-    if(zone.getPoints().size() < ZoneRecord.MIN_POINTS || zone.getPoints().size() > ZoneRecord.MAX_POINTS) {
-      deleteZone(zone.getId());
-      geometryChange();
-      return new ZoneRecord(-1, "");
-    }
+    if(zoneExists(zone.getId())) {
+      String sql = "UPDATE zone SET geometry = GeomFromText('POLYGON((";
 
-    // Update the points table.
-    dbHelper.exec("DELETE FROM point WHERE zone_id = '" + zone.getId() + "'");
-    for(PointRecord point : zone.getPoints())
-      addPoint(point, zone.getId());
-
-    // Update the zone's geometry and label.
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT id FROM zone WHERE id = '" + zone.getId() + "'");
-    if(zoneRecords.moveToNext()) {
-      sql = "UPDATE zone SET geometry = GeomFromText('POLYGON((";
       for(int i = 0; i < zone.getPoints().size(); i++) {
         if(i != 0)
           sql += ",";
         sql += zone.getPoints().get(i).getX() + " " + zone.getPoints().get(i).getY();
       }
       sql += "," + zone.getPoints().get(0).getX() + " " + zone.getPoints().get(0).getY();
-      dbHelper.exec(sql + "))', 4326) WHERE id = '" + zone.getId() + "'");
-      dbHelper.exec("UPDATE zone SET label = '" + DatabaseHelper.escapeString(zone.getLabel()) +
-                    "' WHERE id = '" + zone.getId() + "'");
+
+      dbHelper.exec(sql + "))', 4326) WHERE _id = '" + zone.getId() + "'");
+      dbHelper.exec("UPDATE zone SET label = '" + DatabaseHelper.escapeString(zone.getLabel()) + "' " +
+                    "WHERE _id = '" + zone.getId() + "'");
+
+      dbHelper.exec("DELETE FROM point WHERE zone_id = '" + zone.getId() + "'");
+      for(PointRecord point : zone.getPoints())
+        addPoint(point, zone.getId());
     }
     else
       zone = new ZoneRecord(-1, "");
 
-    zoneRecords.close();
     geometryChange();
     return zone;
   }
 
   public ZoneRecord getZone(int zone_id) {
     ZoneRecord zone = new ZoneRecord(-1, "");
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT id, label FROM zone WHERE id = '" + zone_id + "'");
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT _id, label FROM zone WHERE _id = '" + zone_id + "'");
 
     if(zoneRecords.moveToNext()) {
       zone = new ZoneRecord(zoneRecords.getInt(0), DatabaseHelper.unescapeString(zoneRecords.getString(1)));
-      SpatialCursor pointRecords = dbHelper.prepare("SELECT id, latitude, longitude FROM point WHERE " +
+      SpatialCursor pointRecords = dbHelper.prepare("SELECT _id, latitude, longitude FROM point WHERE " +
                                                       "zone_id = '" + zoneRecords.getInt(0) + "'");
 
       while(pointRecords.moveToNext())
@@ -227,13 +229,104 @@ public class ZoneDatabase {
 
   public List<ZoneRecord> getZones() {
     List<ZoneRecord> zoneList = new LinkedList<ZoneRecord>();
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT id FROM zone");
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT _id FROM zone");
 
     while(zoneRecords.moveToNext())
       zoneList.add(getZone(zoneRecords.getInt(0)));
 
     zoneRecords.close();
     return zoneList;
+  }
+
+  public List<ZoneRecord> getZonesContainingPoint(PointRecord point) {
+    List<ZoneRecord> zoneList = new ArrayList<ZoneRecord>();
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT _id FROM zone WHERE " +
+                                                   "ST_WITHIN(GeomFromText('" +
+                                                     "POINT(" + point.getX() + " " + point.getY() + ")'" +
+                                                     ", 4326), geometry) " +
+                                                 "ORDER BY ST_AREA(geometry) ASC");
+    while(zoneRecords.moveToNext())
+      zoneList.add(getZone(zoneRecords.getInt(0)));
+
+    zoneRecords.close();
+    return zoneList;
+  }
+
+  public List<ZoneRecord> getZonesIntersecting(ZoneRecord zone) {
+    List<ZoneRecord> zoneList = new ArrayList<ZoneRecord>();
+    String sql = "SELECT _id FROM zone WHERE ST_Intersects(geometry, GeomFromText('POLYGON((";
+
+    for(int i = 0; i < zone.getPoints().size(); i++) {
+      if(i != 0)
+        sql += ",";
+      sql += zone.getPoints().get(i).getX() + " " + zone.getPoints().get(i).getY();
+    }
+    sql += "," + zone.getPoints().get(0).getX() + " " + zone.getPoints().get(0).getY();
+
+    SpatialCursor zoneRecords = dbHelper.prepare(sql + "))', 4326)) ORDER BY ST_AREA(geometry) DESC");
+    while(zoneRecords.moveToNext())
+      zoneList.add(getZone(zoneRecords.getInt(0)));
+
+    zoneRecords.close();
+    return zoneList;
+  }
+
+  public PointRecord[] getZoneBounds(int zone_id) {
+    PointRecord[] pointBounds = {new PointRecord(-1, -1, -1, -1), new PointRecord(-1, -1, -1, -1)};
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT " +
+                                                   "MbrMinX(geometry), " +
+                                                   "MbrMinY(geometry), " +
+                                                   "MbrMaxX(geometry), " +
+                                                   "MbrMaxY(geometry) " +
+                                                 "FROM zone WHERE _id = '" + zone_id + "'");
+    if(zoneRecords.moveToNext()) {
+      pointBounds[0] = new PointRecord(-1, -1, zoneRecords.getDouble(0), zoneRecords.getDouble(1));
+      pointBounds[1] = new PointRecord(-1, -1, zoneRecords.getDouble(2), zoneRecords.getDouble(3));
+    }
+
+    zoneRecords.close();
+    return pointBounds;
+  }
+
+  public double distanceBetween(PointRecord point, int zone_id) {
+    double distance = -1;
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT " +
+                                                   "Distance(" +
+                                                   "GeomFromText('POINT(" + point.getX() + " " + point.getY() + ")', 4326), " +
+                                                     "geometry, " +
+                                                     "0" +
+                                                   ") AS dist " +
+                                                 "FROM zone " +
+                                                   "WHERE _id = '" + zone_id + "' AND dist > 0.0 " +
+                                                 "ORDER BY dist ASC LIMIT 1");
+    if(zoneRecords.moveToNext())
+      distance = zoneRecords.getDouble(0);
+
+    zoneRecords.close();
+    return distance;
+  }
+
+  public double distanceToClosestZone(PointRecord point) {
+    double distance = -1;
+    SpatialCursor zoneRecords = dbHelper.prepare("SELECT " +
+                                                 "Distance(" +
+                                                   "GeomFromText('" +
+                                                     "POINT(" +
+                                                       point.getX() + " " +
+                                                       point.getY() +
+                                                     ")', " +
+                                                   "4326), " +
+                                                   "geometry, " +
+                                                   "0" +
+                                                 ") AS dist " +
+                                                 "FROM zone " +
+                                                 "WHERE dist > 0.0 " +
+                                                 "ORDER BY dist ASC LIMIT 1");
+    if(zoneRecords.moveToNext())
+      distance = zoneRecords.getDouble(0);
+
+    zoneRecords.close();
+    return distance;
   }
 
   public void setZoneOccupancy(int zone_id, boolean occupy) {
@@ -296,95 +389,4 @@ public class ZoneDatabase {
     return zoneList;
   }
 
-  public List<ZoneRecord> getZonesContainingPoint(PointRecord point) {
-    List<ZoneRecord> zoneList = new ArrayList<ZoneRecord>();
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT id FROM zone WHERE " +
-                                                   "ST_WITHIN(GeomFromText('" +
-                                                     "POINT(" + point.getX() + " " + point.getY() + ")'" +
-                                                     ", 4326), geometry) " +
-                                                 "ORDER BY ST_AREA(geometry) ASC");
-    while(zoneRecords.moveToNext())
-      zoneList.add(getZone(zoneRecords.getInt(0)));
-
-    zoneRecords.close();
-    return zoneList;
-  }
-
-  public List<ZoneRecord> getZonesIntersecting(ZoneRecord zone) {
-    String sql;
-    List<ZoneRecord> zoneList = new ArrayList<ZoneRecord>();
-
-    sql = "SELECT id FROM zone WHERE ST_Intersects(geometry, GeomFromText('POLYGON((";
-    for(int i = 0; i < zone.getPoints().size(); i++) {
-      if(i != 0)
-        sql += ",";
-      sql += zone.getPoints().get(i).getX() + " " + zone.getPoints().get(i).getY();
-    }
-    sql += "," + zone.getPoints().get(0).getX() + " " + zone.getPoints().get(0).getY();
-
-    SpatialCursor zoneRecords = dbHelper.prepare(sql + "))', 4326)) ORDER BY ST_AREA(geometry) DESC");
-    while(zoneRecords.moveToNext())
-      zoneList.add(getZone(zoneRecords.getInt(0)));
-
-    zoneRecords.close();
-    return zoneList;
-  }
-
-  public PointRecord[] getZoneBounds(int zone_id) {
-    PointRecord[] pointBounds = {new PointRecord(-1, -1, -1), new PointRecord(-1, -1, -1)};
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT " +
-                                                   "MbrMinX(geometry), " +
-                                                   "MbrMinY(geometry), " +
-                                                   "MbrMaxX(geometry), " +
-                                                   "MbrMaxY(geometry) " +
-                                                 "FROM zone WHERE id = '" + zone_id + "'");
-    if(zoneRecords.moveToNext()) {
-      pointBounds[0] = new PointRecord(-1, zoneRecords.getDouble(0), zoneRecords.getDouble(1));
-      pointBounds[1] = new PointRecord(-1, zoneRecords.getDouble(2), zoneRecords.getDouble(3));
-    }
-
-    zoneRecords.close();
-    return pointBounds;
-  }
-
-  public double distanceBetween(PointRecord point, int zone_id) {
-    double distance = -1;
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT " +
-                                                   "Distance(" +
-                                                   "GeomFromText('POINT(" + point.getX() + " " + point.getY() + ")', 4326), " +
-                                                     "geometry, " +
-                                                     "0" +
-                                                   ") AS dist " +
-                                                 "FROM zone " +
-                                                   "WHERE id = '" + zone_id + "' AND dist > 0.0 " +
-                                                 "ORDER BY dist ASC LIMIT 1");
-    if(zoneRecords.moveToNext())
-      distance = zoneRecords.getDouble(0);
-
-    zoneRecords.close();
-    return distance;
-  }
-
-  public double distanceToClosestZone(PointRecord point) {
-    double distance = -1;
-    SpatialCursor zoneRecords = dbHelper.prepare("SELECT " +
-                                                 "Distance(" +
-                                                   "GeomFromText('" +
-                                                     "POINT(" +
-                                                       point.getX() + " " +
-                                                       point.getY() +
-                                                     ")', " +
-                                                   "4326), " +
-                                                   "geometry, " +
-                                                   "0" +
-                                                 ") AS dist " +
-                                                 "FROM zone " +
-                                                 "WHERE dist > 0.0 " +
-                                                 "ORDER BY dist ASC LIMIT 1");
-    if(zoneRecords.moveToNext())
-      distance = zoneRecords.getDouble(0);
-
-    zoneRecords.close();
-    return distance;
-  }
 }
